renaming svn-specific files
